move fix_urls to corehq.util
